code best practise fix
