+++ apps/deeplearning_apps/style_transfer/.gitignore
@@ -1,4 +1,3 @@
-*.pyc
 *.state
 venv/
 *__pycache__/
