+++ apps/deeplearning_apps/style_transfer/app.py
@@ -1,8 +1,7 @@
-from html_content import html
+
 import logging
 
-from h2o_wave import Q, main, app, handle_on, on, ui
-from constants import * 
+from h2o_wave import Q, main, app, handle_on, on
 import os
 
 import cards
@@ -10,6 +9,10 @@ from style import load_model, stylize
 
 # Set up logging
 logging.basicConfig(format='%(levelname)s:\t[%(asctime)s]\t%(message)s', level=logging.INFO)
+
+models_choice = ['candy', 'mosaic', 'rain_princess', 'udnie','fire','black_rainbow','hex']
+source_image_choice = ['house.jpg','beth.jpeg']
+models_images = ['candy.jpg', 'mosaic.jpg', 'rain_princess.jpg', 'udnie.jpg','fire.jpg','black_rainbow.jpg','hex.jpg']
 
 
 @app('/')
@@ -93,50 +96,8 @@ async def serve(q: Q):
         await show_error(q, error=str(error))
 
 async def dashboard_page(q, details=None):
-	models = [ui.choice(name=i, label=i)
-            for i in models_choice]
-	sources = [ui.choice(name=i, label=i)
-            for i in source_image_choice]
-	image_form = [ui.button(name='try_your_image', label='Try Existing Image' if q.user.try_your_image else 'Try Your Image')]
-	if q.user.try_your_image:
-		image_form = image_form + [ui.file_upload(
-            name='upload_image', label='Upload your image', compact=True),]
-	else:
-		image_form = image_form + [
-			ui.dropdown(trigger = True, popup='always', name = 'source_img', label = 'Source Image', value =q.args.source_img or source_image_choice[0], choices=sources),
-		]
-	image_form = image_form + [
-		ui.dropdown(trigger = True, name='style_model', label='Style Model',
-		            value=q.args.style_model or models_choice[0], choices=models),
-        ui.button(name='apply_style', label="Apply Style"),
-        ui.expander(name='preview', expanded=True, label='Style Preview', items=[
-			ui.text("<img src='"+ q.user.template_image_path+"' width='99%' >")
-		]),
-	]
-	# BASE_URL = 'https://mystique-transfer-learning.herokuapp.com'
-	BASE_URL = 'http://localhost:10101'
-	local_path_color, = await q.site.upload([q.user.input_image])
-	local_path_bw, = await q.site.upload([q.user.output_image if q.user.apply_style else q.user.input_image])
-	image_slider_html = html
-
-	image_slider_html = image_slider_html.format(
-		color=BASE_URL + local_path_color, bw=BASE_URL + local_path_bw)
-
-
-	cfg = {
-            'tag': 'dashboard',
-			'image_form': image_form,
-			'image_slider_html': image_slider_html
-    }
-	await render_template(q, cfg)
-	
-async def render_template(q: Q, page_cfg):
-    q.page['main_left'] = ui.form_card(box=ui.box(
-        zone='main', width='20%', order=1), title='', items=page_cfg['image_form'])
-    q.page['main_right'] = ui.frame_card(box=ui.box(
-            zone='main', width='80%', order=2), title='', content=page_cfg['image_slider_html'])
-    await q.page.save()
-
+    cfg = await cards.create_dashboard(q,models_choice,source_image_choice)
+    await cards.render_template(q, cfg)
 
 async def initialize_app(q: Q):
     """
