--- apps/deeplearning_apps/style_transfer/constants.py
@@ -1,4 +1,0 @@
-	
-models_choice = ['candy', 'mosaic', 'rain_princess', 'udnie','fire','black_rainbow','hex']
-source_image_choice = ['house.jpg','beth.jpeg']
-models_images = ['candy.jpg', 'mosaic.jpg', 'rain_princess.jpg', 'udnie.jpg','fire.jpg','black_rainbow.jpg','hex.jpg']
